SentimentAnalysis/fastapi/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
pyyaml